refactor(client): split admin notification components
