+++ client/src/hooks/useAdminWebPush.js
@@ -14,86 +14,15 @@ import {
   registerAdminPushSubscription,
 } from "@/services/adminNotificationsService";
 
-const WEB_PUSH_OPERATION_TIMEOUT_MS = 10000;
-const WEB_PUSH_TIMEOUT_ERROR = "admin-web-push-timeout";
-const WEB_PUSH_SERVICE_WORKER_UNAVAILABLE_ERROR = "admin-web-push-service-worker-unavailable";
-
-function urlBase64ToUint8Array(base64String) {
-  const padding = "=".repeat((4 - (base64String.length % 4)) % 4);
-  const base64 = `${base64String}${padding}`.replace(/-/g, "+").replace(/_/g, "/");
-  const rawData = window.atob(base64);
-  return Uint8Array.from([...rawData].map((character) => character.charCodeAt(0)));
-}
-
-function getSerializedSubscription(subscription) {
-  const serialized = subscription.toJSON();
-  return {
-    endpoint: serialized.endpoint,
-    keys: {
-      p256dh: serialized.keys.p256dh,
-      auth: serialized.keys.auth,
-    },
-    userAgent: navigator.userAgent,
-  };
-}
-
-async function summarizeEndpoint(endpoint) {
-  const cryptoSubtle = globalThis.crypto?.subtle;
-  if (!endpoint || !cryptoSubtle || typeof TextEncoder === "undefined") {
-    return null;
-  }
-
-  let endpointHost = "invalid";
-  try {
-    endpointHost = new URL(endpoint).host;
-  } catch {}
-
-  const endpointBytes = new TextEncoder().encode(endpoint);
-  const endpointHashBuffer = await cryptoSubtle.digest("SHA-256", endpointBytes);
-  const endpointHash = [...new Uint8Array(endpointHashBuffer)]
-    .map((byte) => byte.toString(16).padStart(2, "0"))
-    .join("")
-    .slice(0, 12);
-
-  return { endpointHost, endpointHash };
-}
-
-function hasWebPushSupport() {
-  return (
-    typeof window !== "undefined" &&
-    "Notification" in window &&
-    "serviceWorker" in navigator &&
-    "PushManager" in window
-  );
-}
-
-function withWebPushTimeout(promise) {
-  let timeoutId;
-  const timeoutPromise = new Promise((_, reject) => {
-    timeoutId = setTimeout(() => reject(new Error(WEB_PUSH_TIMEOUT_ERROR)), WEB_PUSH_OPERATION_TIMEOUT_MS);
-  });
-
-  return Promise.race([promise, timeoutPromise]).finally(() => clearTimeout(timeoutId));
-}
-
-function getWebPushFailureReason(error) {
-  if (error?.message === "admin-web-push-unavailable") return "vapidUnavailable";
-  if (error?.message === WEB_PUSH_SERVICE_WORKER_UNAVAILABLE_ERROR) return "serviceWorkerUnavailable";
-  if (error?.message === WEB_PUSH_TIMEOUT_ERROR) return "timeout";
-  return "failed";
-}
-
-async function getReadyServiceWorkerRegistration() {
-  try {
-    return await withWebPushTimeout(navigator.serviceWorker.ready);
-  } catch (error) {
-    const existingRegistration = await navigator.serviceWorker.getRegistration?.();
-    if (!existingRegistration) {
-      throw new Error(WEB_PUSH_SERVICE_WORKER_UNAVAILABLE_ERROR);
-    }
-    throw error;
-  }
-}
+import {
+  getReadyServiceWorkerRegistration,
+  getSerializedSubscription,
+  getWebPushFailureReason,
+  hasWebPushSupport,
+  summarizeEndpoint,
+  urlBase64ToUint8Array,
+  withWebPushTimeout,
+} from "./adminWebPushUtils";
 
 export function useAdminWebPush() {
   const isSupported = useMemo(() => hasWebPushSupport(), []);
